requirements.txt
aiomysql==0.2.0
asttokens==2.4.0
backcall==0.2.0
blinker==1.6.2
cffi==1.15.1
click==8.1.7
cryptography==41.0.3
databases==0.8.0
decorator==5.1.1
executing==1.2.0
Flask==2.3.3
greenlet==2.0.2
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
parso==0.8.3
peewee==3.16.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
PyMySQL==1.1.0
redis==5.0.0
six==1.16.0
SQLAlchemy==1.4.49
stack-data==0.6.2
traitlets==5.10.0
typesystem==0.3.1
wcwidth==0.2.6
Werkzeug==2.3.7
wgconfig==1.0.0
